requirements.txt
numpy
onlinehd
pandas
Requests
scikit_learn
torch
tqdm
rich
pyserial
matplotlib